add fleet, osc relay, cicd build step
