--- components/python/fleet/.gitignore
@@ -1,2 +1,0 @@
-build/
-dist/
