--- components/python/stagebridge/x86_64.spec
@@ -1,44 +1,0 @@
-# -*- mode: python ; coding: utf-8 -*-
-
-
-a = Analysis(
-    ['main.py'],
-    pathex=[],
-    binaries=[],
-    datas=[('templates', 'templates')],
-    hiddenimports=['mido.backends.rtmidi'],
-    hookspath=[],
-    hooksconfig={},
-    runtime_hooks=[],
-    excludes=[],
-    noarchive=False,
-    optimize=0,
-)
-pyz = PYZ(a.pure)
-
-exe = EXE(
-    pyz,
-    a.scripts,
-    a.binaries,
-    a.datas,
-    [],
-    name='stagebridge',
-    debug=False,
-    bootloader_ignore_signals=False,
-    strip=False,
-    upx=True,
-    upx_exclude=[],
-    runtime_tmpdir=None,
-    console=False,
-    disable_windowed_traceback=False,
-    argv_emulation=False,
-    target_arch='x86_64',
-    codesign_identity=None,
-    entitlements_file=None,
-)
-app = BUNDLE(
-    exe,
-    name='Stagebridge.app',
-    icon=None,
-    bundle_identifier='com.trekkerholdings.stagebridge',
-)
