--- docker-compose.yml
@@ -1,12 +1,0 @@
-version: "3.8"
-services:
-  stagebridge:
-    build: .
-    ports:
-      - "3000:3000"
-      - "3001:3001"
-      - "8000:8000"
-      - "9000:9000"
-    devices:
-      - /dev/snd:/dev/snd
-    # privileged: true # only if needed
