+++ extras/generator/package.json
