+++ extras/generator/public/main.js
