+++ extras/generator/public/style.css
