+++ extras/generator/src/components/GeneratorPage.jsx
