+++ extras/generator/src/components/Layout.jsx
