+++ extras/generator/vite.config.js
