+++ stagebridge/osc_server.py
@@ -1,4 +1,3 @@
-# osc_server.py
 import threading
 import time
 from pythonosc.dispatcher import Dispatcher
@@ -6,35 +5,62 @@ from pythonosc.osc_server import ThreadingOSCUDPServer
 from mido import Message
 import shared_state
 
+def _relay_to_discovered_devices(address, *args):
+    """Relays OSC message to all discovered StageBridge devices."""
+    if not shared_state.discovered_devices:
+        print("No discovered devices to relay to")
+        return
+    
+    print(f"Relaying OSC message {address} {args} to {len(shared_state.discovered_devices)} devices")
+    
+    for device_name, device_info in shared_state.discovered_devices.items():
+        try:
+            device_info['client'].send_message(address, args)
+            print(f"  -> Relayed to {device_info['name']} ({device_info['ip']})")
+        except Exception as e:
+            print(f"  -> Failed to relay to {device_info['name']}: {e}")
+
 def _osc_handler(address, *args):
     """Handles incoming OSC messages and translates them to physical MIDI."""
     print(f"OSC Received: {address} {args}")
-    if not shared_state.midi_out_port:
-        print("Warning: MIDI output port not configured or open.")
-        return
     
+    # Check for predefined mappings first
+    mapping_found = False
     for mapping in shared_state.config.get("osc_mappings", []):
         if mapping["osc_address"] == address:
+            mapping_found = True
+            
+            if not shared_state.midi_out_port:
+                print("Warning: MIDI output port not configured or open.")
+                continue
+            
             midi_sequence = mapping.get("midi_sequence", [])
-            if not midi_sequence: continue
+            if not midi_sequence: 
+                continue
             
             print(f"Found mapping for {address}. Sending sequence...")
             for midi_info in midi_sequence:
                 try:
                     msg_type = midi_info["type"]
-                    # channel = int(midi_info["channel"]) - 1
                     channel = int(midi_info["channel"])
+                    
                     if msg_type == "program_change":
                         msg = Message("program_change", channel=channel, program=int(midi_info["program"]))
                     elif msg_type == "control_change":
                         msg = Message("control_change", channel=channel, control=int(midi_info["control"]), value=int(midi_info["value"]))
-                    else: continue
+                    else: 
+                        continue
                     
                     print(f"  -> Sending MIDI: {msg}")
                     shared_state.midi_out_port.send(msg)
-                    time.sleep(0.01) # Small delay for reliability
+                    time.sleep(0.01)
                 except Exception as e:
                     print(f"Error processing step in sequence for {address}: {e}")
+    
+    # If no mapping found, relay to discovered devices
+    if not mapping_found:
+        print(f"No mapping found for {address}, relaying to discovered devices")
+        _relay_to_discovered_devices(address, *args)
 
 def start_osc_server():
     """Starts the OSC server in a background thread."""
