+++ stagebridge/templates/fleet/client.js
