+++ stagebridge/templates/public/app.js
